+++ docs/natbib/formatters/nodes.py
@@ -310,6 +310,7 @@ call = Call()
 emph = Emph()
 field = Field()
 join = Join()
+concat = Join()
 optional = Optional()
 text = Text()
 boolean = Boolean()
